Move #GameSelectorPanel, turns out it is "owned" by #MainPanel. (Also included some cleanup/javadoc changes)
